requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
pillow==9.5.0